+++ esperanto-tools-libs/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-20002] Avoid running exit handlers on forked processes
 - [SW-20203] Remove duplicate profiling of kernel launches
 - [SW-20211] Improve maximum arguments size supported by allowing to clobber the exception buffer
+- [SW-20546] Fix IdentifyThread tracing events without a thread name
 
 ## [0.13.1]
 ### Fixed